Make pg install more idempotent

+++ chef_postgres/recipes/master.rb
@@ -20,6 +20,7 @@ admin_is_generated = node["chef_postgres"]["vars"]["admin_is_generated"]
 repl_user = node["chef_postgres"]["vars"]["repl_user"]
 repl_pass = node["chef_postgres"]["vars"]["repl_pass"]
 backup_dir = node["chef_postgres"]["pg_config"]["backup_directory"]
+user_created = node["chef_postgres"]["vars"]["user_created"]
 
 # Build this on the master so the standbys have the right settings
 template "recovery_conf.source" do
@@ -32,6 +33,7 @@ template "recovery_conf.source" do
                         password: repl_pass,
                         hostname: node["ec2"]["local_hostname"],
                         standby: node["chef_postgres"]["pg_config"]["cluster_type"] } })
+  not_if { user_created }
 end
 
 bash "move_data_directory" do
@@ -55,12 +57,14 @@ bash "move_data_directory" do
   not_if { ::File.exist?("#{node["chef_postgres"]["pg_config"]["data_directory"]}/PG_VERSION") }
   only_if { node["chef_postgres"]["pg_config"]["data_directory_on_separate_drive"] }
   user "postgres"
+  not_if { user_created }
   notifies :run, "ruby_block[log_move_data_directory]", :before
 end
 
 service "start_postgres" do
   action :start
   service_name "postgresql"
+  not_if { user_created }
   notifies :run, "ruby_block[log_start_pg]", :before
 end
 
@@ -71,18 +75,8 @@ bash "create_admin_user" do
     echo "ALTER USER postgres WITH PASSWORD '#{pg_pass}';" | psql -U postgres -d postgres
     echo "CREATE USER #{admin_user} WITH PASSWORD '#{admin_pass}' SUPERUSER CREATEDB CREATEROLE; CREATE DATABASE #{admin_user} OWNER #{admin_user};" | psql -U postgres -d postgres
   EOF_CAU
+  not_if { user_created }
   notifies :run, "ruby_block[log_create_admin]", :before
-end
-
-# Only run this, if generating the info through the defaults.
-file "record_admin" do
-  content "user: #{admin_user} password: #{admin_pass}"
-  group "root"
-  mode "0400"
-  owner "root"
-  path "/etc/postgresql/#{version}/main/admin_login"
-  action :create
-  only_if { admin_is_generated }
 end
 
 bash "create_repl_user" do
@@ -90,7 +84,7 @@ bash "create_repl_user" do
   code <<-EOF_CRU
   echo "CREATE USER #{repl_user} WITH PASSWORD '#{repl_pass}' REPLICATION LOGIN CONNECTION LIMIT 4;" | psql -U postgres -d postgres
   EOF_CRU
-  action :run
+  not_if { user_created }
   notifies :run, "ruby_block[log_create_repl]", :before
 end
 
@@ -98,8 +92,9 @@ bash "create_base_backup" do
   code <<-EOF_CBB
   rm -rf #{backup_dir}/base_backup/*
   pg_basebackup -d 'host=localhost user=#{repl_user} password=#{repl_pass}' -D #{backup_dir}/base_backup --xlog-method=stream
-  tar -C #{backup_dir} -czf #{backup_dir}/base_backup.tgz #{backup_dir}/base_backup/  
+  tar -C #{backup_dir} -czf #{backup_dir}/base_backup.tgz #{backup_dir}/base_backup/
   EOF_CBB
+  not_if { user_created }
   action :run
   notifies :run, "ruby_block[log_create_base_backup]", :before
 end
@@ -110,6 +105,7 @@ ruby_block "s3_upload_backup" do
       { bucket: node["chef_postgres"]["s3"]["bucket"],
         source: "#{backup_dir}/base_backup.tgz" })
   end
+  not_if { user_created }
 end
 
 cookbook_file "Copy backup file" do
@@ -118,17 +114,31 @@ cookbook_file "Copy backup file" do
   owner "ubuntu"
   path "#{backup_dir}/pg_backup.sh"
   source "pg_backup.sh"
+  not_if { user_created }
 end
 
-cron "pg_backup.sh" do 
-  minute "0"
+cron "pg_backup.sh" do
+  minute "2"
   hour "7"
   user "ubuntu"
   command "#{backup_dir}/pg_backup.sh"
+  not_if { user_created }
 end
 
 service "restart_postgres" do
   action :restart
   service_name "postgresql"
   notifies :run, "ruby_block[log_restart_pg]", :before
+  not_if { user_created }
+end
+
+# Do last. If everything works, this keeps things idempotent.
+file "record_admin" do
+  content "admin user: #{admin_user} password: #{admin_pass} \n repl user: #{repl_user} password: #{repl_pass}"
+  group "root"
+  mode "0400"
+  owner "root"
+  path node["chef_postgres"]["vars"]["admin_login_path"]
+  not_if { user_created }
+  action :create
 end
